+++ src/games/stendhal/server/entity/player/Player.java
@@ -193,7 +193,7 @@ public class Player extends DressedEntity implements UseListener {
 		// define outfit
 		Outfit outfit = null;
 		if (template != null && template.has("outfit")) {
-			outfit = new Outfit(template.getInt("outfit"));
+			outfit = new Outfit(template.getLong("outfit"));
 		}
 
 		if (outfit == null || !outfit.isChoosableByPlayers()) {
